// exercises/1.9/basic-ts-api-boilerplate/types.ts
interface Pizza {
  id: number;
  title: string;
  content: string;
}

interface PizzaToUpdate {
  title?: string;
  content?: string;
}

type NewPizza = Omit<Pizza, "id">;

interface Text {
  id: string;
  content: string;
  level: 'easy' | 'medium' | 'hard';
}

type NewText = Omit<Text, "id">;


export type { Pizza, NewPizza, PizzaToUpdate, Text, NewText };